added documentation to non-test code
