--- src/main/java/screens/CreateAccount/CreateAccount/AccountCreationFailed.java
@@ -1,7 +1,0 @@
-package screens.CreateAccount.CreateAccount;
-
-public class AccountCreationFailed extends RuntimeException{
-    public AccountCreationFailed(String error) {
-        super(error);
-    }
-}
