+++ src/main/java/useCases/CreateAccount/CreateAccountPresenter.java
@@ -1,5 +1,8 @@
 package useCases.CreateAccount;
 
+/**
+ * An interface which hides the details of creating a viewModel for CreateAccount purposes
+ */
 public interface CreateAccountPresenter {
     CreateAccountResponseModel prepareSuccessView(CreateAccountResponseModel success);
